requirements.txt
requests==2.23.0
opencv_contrib_python==4.1.2.30
scipy==1.4.1
keras==2.6.0
pandas==1.1.5
torch==1.9.0+cu102
Flask==1.1.4
tensorflow==2.6.0
imutils==0.5.4
numpy==1.19.5
gradio==2.1.0
Pillow==8.3.0
filterpy==1.4.5
sktime==0.8.0
sympy==1.9
scikit-image==0.18.3
protobuf<=3.20.1